Move DEBUG code paths from the main `SceneDelegate` to a new `DebuggingSceneDelegate`
